Implemented a basic sorting in CUDA cmopilation of Quicksilver.
Particles are sorted by CUDA block, localy and in shared memory
so that collision events are always sorted to the front of a block.
Occurs after every call to MC_Segment_Outcome.
Additionally after processing particles they are again sorted to
place all particles that need to keepTracking in the front of the
block, keeping active particles together.